Revert some changes that broke things.
The check for OS is not needed as it's the binary off ffmpeg that is relevant.

The fallback from ffprobe to ffprobe makes no sense (fallback to ffmpeg did) - remove the fallback entirely.

Convert_path seemed a bit ass as indentation didn't make sense so couldn't follow the logic.
Will try to re-introduce as it's a good idea to avoid cygpath.