Version 3
Prompt:-

Please ensure that blank lines do not contain any spaces. (i.e. keep the blank line but make it actually blank). Is there any code that is redundant or not used?
As well as showing the frame numbers after iteration, could you also add the gaps between the frames in this output in a readable way (also all on one line)?